Replace the View behaviour with a number of protocols
This removes the View behavior completely, sacrificing some minor features
(pagination and top level meta and links handling), and delegates all resource
related handling to a number of protocols. It serves as an experiment to test
wether a protocol-based approach is beneficial while allowing for similar ease
of implementation for the library user.